refactor: move image utils
